moved checkValidFunction to own file within utils folder
